+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Enhancements
     - Looking of the data table is now done via a named lookup table instead
       of calling the database state server
-    - Old data tables are deleted with a 30 second delay after reloading
+    - Old data tables are deleted with a configurable delay after reloading
       to avoid race conditions (and the resulting empty lookup responses)
 
 ## v0.18.0 (2017-12-31)